chore: rename *_deps.ts -> deps.*.ts
As per David's suggestion.
