+++ docs/getting-started/adding-interactivity.md
@@ -31,7 +31,7 @@ Here is an example of an island component that counts down to a specific time.
 // islands/Countdown.tsx
 
 /** @jsx h */
-import { h, useEffect, useMemo, useState } from "../client_deps.ts";
+import { h, useEffect, useMemo, useState } from "../deps.client.ts";
 
 const timeFmt = new Intl.RelativeTimeFormat("en-US");
 
@@ -73,7 +73,7 @@ client with the correct props:
 // routes/countdown.tsx
 
 /** @jsx h */
-import { h } from "../client_deps.ts";
+import { h } from "../deps.client.ts";
 import { Countdown } from "../islands/Countdown.tsx";
 
 export default function Page() {
