+++ tests/fixture/deps.client.ts
