+++ tests/fixture/islands/Test.tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { h } from "../client_deps.ts";
+import { h } from "../deps.client.ts";
 
 export default function Test(props: { message: string }) {
   return (
