+++ tests/fixture/routes/[name].tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { h } from "../client_deps.ts";
+import { h } from "../deps.client.ts";
 
 interface Props {
   params: Record<string, string | string[]>;
