+++ tests/fixture/routes/_500.tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { ErrorPageProps, h } from "../client_deps.ts";
+import { ErrorPageProps, h } from "../deps.client.ts";
 
 export default function Error500Page({ error }: ErrorPageProps) {
   return <p>500 internal error: {(error as Error).message}</p>;
