+++ tests/fixture/routes/books/[id].tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { h, PageProps, RouteConfig } from "../../client_deps.ts";
+import { h, PageProps, RouteConfig } from "../../deps.client.ts";
 
 export default function Page(props: PageProps) {
   return <div>Book {props.params.id}</div>;
