+++ tests/fixture/routes/props/[id].tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { h, PageProps } from "../../client_deps.ts";
+import { h, PageProps } from "../../deps.client.ts";
 
 export default function Home(props: PageProps) {
   return <div>{JSON.stringify(props)}</div>;
