+++ tests/fixture/routes/static.tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { asset, h } from "../client_deps.ts";
+import { asset, h } from "../deps.client.ts";
 
 export default function StaticPage() {
   return (
