+++ tests/fixture_error/main.ts
@@ -4,7 +4,7 @@
 /// <reference lib="deno.ns" />
 /// <reference lib="deno.unstable" />
 
-import { start } from "./server_deps.ts";
+import { start } from "./deps.server.ts";
 import manifest from "./fresh.gen.ts";
 
 await start(manifest);
