+++ www/components/Footer.tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { h, tw } from "../client_deps.ts";
+import { h, tw } from "../deps.client.ts";
 
 const LINKS = [
   {
